Additional fixes.

+++ src/main/java/com/kuborros/FurBotNeo/commands/MusicCommands/MusicTimeCommand.java
@@ -62,7 +62,7 @@ public class MusicTimeCommand extends MusicCommand {
             }
             long seconds;
             try {
-                seconds = (min ? 60 : 1) * Integer.parseInt(val);
+                seconds = (min ? 60 : 1) * Long.parseLong(val);
                 long milis = (seconds * 1000);
                 long duration = getPlayer(guild).getPlayingTrack().getDuration();
                 if (duration <= milis) {
